Move `update_tuple_item` to `eth2._utils.tuple`
