+++ eth2/beacon/validator_status_helpers.py
@@ -1,13 +1,6 @@
-from typing import (
-    Any,
-    Iterable,
-    Sequence,
+from eth2._utils.tuple import (
+    update_tuple_item,
 )
-
-from eth_utils import (
-    ValidationError,
-)
-
 from eth2.beacon.enums import (
     ValidatorRegistryDeltaFlag,
     ValidatorStatusFlags,
@@ -23,30 +16,6 @@ from eth2.beacon.typing import (
     SlotNumber,
     ValidatorIndex,
 )
-
-
-#
-# Helper for updating tuple item
-#
-def update_tuple_item(tuple_data: Sequence[Any],
-                      index: int,
-                      new_value: Any) -> Iterable[Any]:
-    """
-    Update the ``index``th item of ``tuple_data`` to ``new_value``
-    """
-    list_data = list(tuple_data)
-
-    try:
-        list_data[index] = new_value
-    except IndexError:
-        raise ValidationError(
-            "the length of the given tuple_data is {}, the given index {} is out of index".format(
-                len(tuple_data),
-                index,
-            )
-        )
-    else:
-        return tuple(list_data)
 
 
 #
